file restructuring
Moved all _lib files into a libs folder
