+++ digital_game_design/complete/src/libs/timer_lib.js
@@ -1,0 +1,13 @@
+export function updateTimer(gameState) {
+    let timeDif = gameState.elapsedTime;
+    if (gameState.playState === 'playing') {
+        if (gameState.startTime) {
+            const curTime = new Date();
+            timeDif = curTime.valueOf() - gameState.startTime.valueOf();
+        } else {
+            gameState.startTime = new Date();
+        }
+    }
+
+    gameState.elapsedTime = timeDif;
+}
